se actualizan las claves primarias de las clases del repositorio

+++ src/main/java/com/hotel/repositorio/FacturaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hotel.modelo.Factura;
 
-public interface FacturaRepositorio extends CrudRepository<Factura, Integer> {
+public interface FacturaRepositorio extends CrudRepository<Factura, Long> {
 
 }
