+++ src/main/java/com/hotel/repositorio/HabitacionRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.hotel.modelo.Habitacion;
 
-public interface HabitacionRepositorio extends CrudRepository<Habitacion, Integer> {
+public interface HabitacionRepositorio extends CrudRepository<Habitacion, Long> {
 
 }
